Split up mds-db into more manageable files.
